chore: Build gcc 2.34 and 2.39
As part of debugging the permissions issues, I left around a 2.34 release from a local build. This lead to the 2.34 build failing due to the release already existing and further canceling the 2.39 build. Deleted the old release and using this to kick off another build of just the ones that didnt complete